cluster-score-be/requirements.txt
fastapi
uvicorn
httpx
pydantic
openai
gspread
oauth2client
python-dotenv
beautifulsoup4      